Move consumers from `pvt` to `JettyConfig`
